requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
aiostream==0.5.2
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
attrs==24.3.0
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dataclasses-json==0.6.7
deprecated==1.2.15
deepeval==0.9.7
dirtyjson==1.0.8
distro==1.9.0
docx2txt==0.8
einops==0.7.0
fastapi==0.109.2
filelock==3.16.1
filetype==1.2.0
flake8==7.1.1
frozenlist==1.5.0
fsspec==2024.12.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.27.1
idna==3.10
instructorembedding==1.0.1
jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonlines==4.0.0
llama-cloud==0.1.9
llama-index-agent-openai==0.4.2
llama-index-cli==0.4.0
llama-index-core==0.12.11
llama-index-embeddings-huggingface==0.4.0
llama-index-embeddings-instructor==0.3.0
llama-index-embeddings-nomic==0.6.0
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.3
llama-index-llms-ollama==0.5.0
llama-index-llms-openai==0.3.13
llama-index-multi-modal-llms-openai==0.4.2
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.4
llama-index-readers-llama-parse==0.4.0
llama-index-vector-stores-postgres==0.4.1
llama-index==0.12.11
llama-parse==0.5.19
loguru==0.7.3
markdown-it-py==3.0.0
markupsafe==3.0.2
marshmallow==3.23.3
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
nomic==3.4.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
ollama==0.4.7
openai==1.59.3
packaging==24.2
pandas==2.2.3
pgvector==0.3.6
pillow==11.1.0
propcache==0.2.1
psycopg2-binary==2.9.10
pyarrow==19.0.0
pycodestyle==2.12.1
pydantic-core==2.23.4
pydantic==2.9.2
pyflakes==3.2.0
pygments==2.19.0
pyjwt==2.10.1
pypdf==5.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
sentence-transformers==2.7.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlalchemy==2.0.36
starlette==0.36.3
striprtf==0.0.26
sympy==1.13.1
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
torch==2.2.1
tqdm==4.67.1
transformers==4.48.3
triton==2.2.0
typing-extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.23.2
uvloop==0.21.0
watchfiles==1.0.3
websockets==14.1
win32-setctime==1.2.0
wrapt==1.17.0
yarl==1.18.3
accelerate>=0.26.0
bitsandbytes==0.45.2
